feat : 장바구니 기능 구현

+++ 05_spa-mall/routes/goods.js
@@ -1,7 +1,28 @@
 const express = require("express");
 const router = express.Router();
-const Goods = require("../schemas/goods")
+const Goods = require("../schemas/goods");
+const authMiddleware = require("../middlewares/auth-middleware");
 
+// 장바구니 조회 api
+router.get("/goods/cart", authMiddleware, async (req, res) => {
+  const { userId } = res.locals.user;
+
+  const carts = await Cart.find({ userId: userId });
+  const goodsIds = carts.map((cart) => {
+    return cart.goodsId;
+  })
+
+  const goods = await Goods.find({ goodsId: goodsIds });
+  const results = carts.map((cart) => {
+    return {
+      "quantity": cart.quantity,
+      "goods": goods.find((item) => item.goodsId === cart.goodsId),
+    }
+  })
+  res.json({
+    "carts": results
+  })
+});
 
 
 // 상품 목록 조회 API
@@ -46,16 +67,17 @@ router.get("/goods/:goodsId", async (req, res) => {
 });
 
 
-
-
-
 const Cart = require("../schemas/cart.js");
 
-router.post("/goods/:goodsId/cart", async (req, res) => {
+
+// 장바구니 등록 API
+router.post("/goods/:goodsId/cart", authMiddleware, async (req, res) => {
+  const { userId } = res.locals.user;
   const { goodsId } = req.params;
   const { quantity } = req.body;
 
-  const existsCarts = await Cart.find({ goodsId });
+  // 사용자정보를 가지고 장바구니 조회
+  const existsCarts = await Cart.find({ userId, goodsId });
   if (existsCarts.length) {
     return res.status(400).json({
       success: false,
@@ -63,35 +85,44 @@ router.post("/goods/:goodsId/cart", async (req, res) => {
     })
   }
 
-  await Cart.create({ goodsId, quantity });
+  // 해당하는 사용자의 정보를 가지고 장바구니에 상품 등록
+  await Cart.create({ userId, goodsId, quantity });
 
   res.json({ result: "success" });
 })
 
-router.put("/goods/:goodsId/cart", async (req, res) => {
+
+// 장바구니 수정 API
+router.put("/goods/:goodsId/cart", authMiddleware, async (req, res) => {
+  const { userId } = res.locals.user;
   const { goodsId } = req.params;
   const { quantity } = req.body;
 
-  const existsCarts = await Cart.find({ goodsId });
-  if (existsCarts.length) {
+  const existsCarts = await Cart.find({ userId, goodsId });
+  if (existsCarts.length > 0) {
     await Cart.updateOne(
-      { goodsId: goodsId },
+      { userId, goodsId: goodsId },
       { $set: { quantity: quantity } }
-    )
+    );
   }
   res.status(200).json({ success: true });
-})
+});
 
-router.delete("/goods/:goodsId/cart", async (req, res) => {
+
+
+// 장바구니 삭제 API
+router.delete("/goods/:goodsId/cart", authMiddleware, async (req, res) => {
+  const { userId } = res.locals.user;
   const { goodsId } = req.params;
 
-  const existsCarts = await Cart.find({ goodsId });
+  const existsCarts = await Cart.find({ userId, goodsId });
   if (existsCarts.length) {
-    await Cart.deleteOne({ goodsId });
+    await Cart.deleteOne({ userId, goodsId });
   }
 
   res.json({ result: "success" });
 })
+
 
 router.post("/goods", async (req, res) => {
   const { goodsId, name, thumbnailUrl, category, price } = req.body;
